requirements.txt
fastapi[standard]==0.114.0
aiogram==3.13.0
pydantic==2.8.2
SQLAlchemy==2.0.34
python-dotenv~=1.0.1